# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Goblin L00t API - Twitch bot + giveaway dashboard
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats/overview:
    get:
      operationId: getStatsOverview
      tags: [stats]
      summary: Get dashboard overview stats
      parameters:
        - $ref: "#/components/parameters/StatsRange"
      responses:
        "200":
          description: Overview statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsOverview"
  /stats/commands:
    get:
      operationId: getCommandStats
      tags: [stats]
      summary: Get command usage statistics
      parameters:
        - $ref: "#/components/parameters/StatsRange"
      responses:
        "200":
          description: Command usage stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommandStat"
  /stats/top-looters:
    get:
      operationId: getTopLooters
      tags: [stats]
      summary: Get top users by loot count
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - $ref: "#/components/parameters/StatsRange"
      responses:
        "200":
          description: Top looters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserStat"
  /stats/engagement:
    get:
      operationId: getEngagementReport
      tags: [stats]
      summary: Get engagement tips and metrics for the selected window
      parameters:
        - $ref: "#/components/parameters/StatsRange"
      responses:
        "200":
          description: Engagement report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngagementReport"
  /stream/status:
    get:
      operationId: getStreamStatus
      tags: [stats]
      summary: Get the caller's current stream session status
      responses:
        "200":
          description: Stream status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamStatus"
        "401":
          description: Unauthorized
  /stream/start:
    post:
      operationId: startStreamSession
      tags: [stats]
      summary: Mark the caller's stream session as started (timestamps now)
      responses:
        "200":
          description: Stream started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamStatus"
        "401":
          description: Unauthorized
  /stream/end:
    post:
      operationId: endStreamSession
      tags: [stats]
      summary: Clear the caller's stream session (sets streamStartedAt to null)
      responses:
        "200":
          description: Stream ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamStatus"
        "401":
          description: Unauthorized
...
components:
  parameters:
    StatsRange:
      name: range
      in: query
      description: Time window for stats. `stream` filters since the caller's streamStartedAt.
      schema:
        type: string
        enum: [day, week, month, year, all, stream]
        default: all
  schemas:
...
    StatsOverview:
      type: object
      properties:
        totalGiveaways:
          type: integer
        activeGiveaway:
          type: boolean
        totalLootDrops:
          type: integer
        totalCommandsUsed:
          type: integer
        uniqueUsers:
          type: integer
        recentEntries:
          type: integer
      required:
        - totalGiveaways
        - activeGiveaway
        - totalLootDrops
        - totalCommandsUsed
        - uniqueUsers
        - recentEntries
    CommandStat:
      type: object
      properties:
        command:
          type: string
        usageCount:
          type: integer
        lastUsedAt:
          type: ["string", "null"]
      required:
        - command
        - usageCount
        - lastUsedAt
    UserStat:
      type: object
      properties:
        username:
          type: string
        lootCount:
          type: integer
        totalPoints:
          type: integer
        bestRarity:
          type: string
      required:
        - username
        - lootCount
        - totalPoints
        - bestRarity
...
    StreamStatus:
      type: object
      properties:
        streamStartedAt:
          type: ["string", "null"]
        isLive:
          type: boolean
      required:
        - streamStartedAt
        - isLive
    EngagementTip:
      type: object
      properties:
        id:
          type: string
        severity:
          type: string
          enum: [info, warn]
        title:
          type: string
        detail:
          type: string
      required: [id, severity, title, detail]
    EngagementReport:
      type: object
      properties:
        range:
          type: string
        since:
          type: ["string", "null"]
        metrics:
          type: object
          properties:
            totalLoot:
              type: integer
            totalCommands:
              type: integer
            totalGiveaways:
              type: integer
            uniqueChatters:
              type: integer
          required: [totalLoot, totalCommands, totalGiveaways, uniqueChatters]
        tips:
          type: array
          items:
            $ref: "#/components/schemas/EngagementTip"
      required: [range, metrics, tips]
...