# lib/api-zod/src/generated/types/streamStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Goblin L00t API - Twitch bot + giveaway dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface StreamStatus {
  /** @nullable */
  streamStartedAt: string | null;
  isLive: boolean;
}